Google_ADK_A2A/a2a-project-main/jeff_agent/pyproject.toml
[project]
name = "jeff-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.25",
    "google-adk>=1.27.4",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.22",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.1.3",
    "python-dotenv>=1.2.2",
    "sse-starlette>=3.3.3",
    "starlette>=0.52.1",
    "uvicorn>=0.42.0",
]
